Refactor code for setting background CSS class on body
Fixing a TODO in the code for setting the weather condition CSS class
on body.  The issue with the existing implementation is that it will
wipe out any/all existing CSS classes on body.  It wasn’t causing an
issue because no other classes were being set, but still not ideal.